Соединит похожие модули в один
