+++ qr.py
@@ -1,6 +1,9 @@
 import qrcode
 import vk_api
 from auth import vk
+from PIL import Image
+from pyzbar.pyzbar import decode
+import requests
 
 
 def qrgen(data):
@@ -15,3 +18,14 @@ def qrgen(data):
     attachment = f'photo{owner_id}_{photo_id}_{access_key}'
     print(attachment)
     return attachment
+
+
+def qrdecode(link):
+    r = requests.get(link)
+    with open('decodeqr.png', 'wb') as f:
+        f.write(r.content)
+    data = decode(Image.open('decodeqr.png'))
+    if data:
+        return data[0].data.decode("utf-8")
+    elif data == []:
+        return ""
